backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
pillow>=10.0.0
sqlmodel>=0.0.8
python-dotenv>=1.0.0
